Implement variant search to support the different analysis modes
